Improved search screen
I replaced the static search screen with a screen that displays a sample
result for any search query. This still needs to be debugged and tested.
I'm getting lots of errors in Eclipse, but I'm not sure if I'm doing
something wrong or if I've just encountered a bug in the IDE.