Improved Buy screen
Began a total rewrite of the Buy screen, replacing old layout with a
ListView. There's still a lot of work left to do here. We may also want
to consider using ListViews in the Sell screen and other screens with
lists, although this might not be the easiest way to make a nicely
formatted, dynamically constructed list.

I also replaced the sample book images with much smaller versions to
prevent crashes. We may need to eventually load images though Java code
rather than XML in order to avoid memory problems.