Partial fix for crash on non-ISBN barcode scan
This fix prevents an exception from occurring if the user scans
something other than an EAN-13 barcode. However, no message is
displayed, because attempting to display a dialog caused an
IllegalStateException.

We still need to look at other possible error states, such as an EAN-13
barcode that doesn't contain an ISBN code.